split test support functions into separate files
